refactor(api): move target-profile-repository into prescription
