+++ api/scripts/prod/insert-missing-pole-emploi-sending-from-date.js
@@ -4,7 +4,6 @@ import dayjs from 'dayjs';
 
 import { disconnect, knex } from '../../db/knex-database-connection.js';
 import * as badgeAcquisitionRepository from '../../lib/infrastructure/repositories/badge-acquisition-repository.js';
-import * as targetProfileRepository from '../../lib/infrastructure/repositories/target-profile-repository.js';
 import * as badgeRepository from '../../src/evaluation/infrastructure/repositories/badge-repository.js';
 import * as userRepository from '../../src/identity-access-management/infrastructure/repositories/user.repository.js';
 import * as campaignRepository from '../../src/prescription/campaign/infrastructure/repositories/campaign-repository.js';
@@ -14,6 +13,7 @@ import * as campaignParticipationRepository from '../../src/prescription/campaig
 import { campaignParticipationResultRepository } from '../../src/prescription/campaign-participation/infrastructure/repositories/campaign-participation-result-repository.js';
 import * as poleEmploiSendingRepository from '../../src/prescription/campaign-participation/infrastructure/repositories/pole-emploi-sending-repository.js';
 import { CampaignParticipationStatuses } from '../../src/prescription/shared/domain/constants.js';
+import * as targetProfileRepository from '../../src/prescription/target-profile/infrastructure/repositories/target-profile-repository.js';
 import { Assessment } from '../../src/shared/domain/models/Assessment.js';
 import * as assessmentRepository from '../../src/shared/infrastructure/repositories/assessment-repository.js';
 import * as organizationRepository from '../../src/shared/infrastructure/repositories/organization-repository.js';
