+++ api/tests/prescription/target-profile/integration/infrastructure/repositories/target-profile-repository_test.js
@@ -1,9 +1,9 @@
 import _ from 'lodash';
 
-import * as targetProfileRepository from '../../../../lib/infrastructure/repositories/target-profile-repository.js';
-import { NotFoundError } from '../../../../src/shared/domain/errors.js';
-import { TargetProfile } from '../../../../src/shared/domain/models/index.js';
-import { catchErr, databaseBuilder, expect } from '../../../test-helper.js';
+import * as targetProfileRepository from '../../../../../../src/prescription/target-profile/infrastructure/repositories/target-profile-repository.js';
+import { NotFoundError } from '../../../../../../src/shared/domain/errors.js';
+import { TargetProfile } from '../../../../../../src/shared/domain/models/index.js';
+import { catchErr, databaseBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Integration | Repository | Target-profile', function () {
   describe('#get', function () {
